Report

+++ database/migrations/2023_05_08_094756_create_transaction_accounts_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('description')->nullable();
-            $table->float('ammount_kredit');
-            $table->float('ammount_debit');
+            $table->decimal('ammount_kredit', 14, 2);
+            $table->decimal('ammount_debit', 14, 2);
             $table->foreignId('accounting_group_id')->constrained();
             $table->timestamps();
         });
